expression parse stack
This time I am actually making the stack as an independent object that
you can just call and get operations out of. This sounds way neater than
what I have done in the past. It crashes when it gets to EOF, when it
should buffer the next token and pop the whole stack. Ah well.