dynamic buffers
This is my spin on Sean's stretchy buffers. Going for a more classic fat
pointer approach, which requires defining more types, and dealing with
more `.data` nonsense, but I think it will be more comfortable for me.

With this I can start building a buffer of operations, to build
functions and so on.