compile array literals into instructions
This was a bigger ordeal than I expected. This is just one out of three
steps, to initialise arrays, and not even use them... Mainly because I
decided to design and start to implement an emplacement based approach,
which requires more fiddling to inject the allocation instruction
earlier in the bytecode than any of the emplaced values themselves, once
the array as a whole has been parsed and type checked.

The expression code is starting to get pretty hairy, and it seems
slightly unnecessary at the moment; I think that it turns out that
pushing constants or variables onto the stack only really matters for
function calls, so it is actually the minority case rather than the
majority case, and so I should go back to my approach in previous
parsers, of storing a stack of refs which may or may not be on the
actual variable stack yet, and then pop from the ref stack instead for
most operations, and then just make a special case of the comma operator
which pushes multi-values ready for function calls or for multiple
assignment/multiple return.

Also the grouping and bracket structure is probably a little bit
complicated right now, and might become simpler after the ref stack
change is made, or after more explicit factoring is done. Also I like
trailing commas a lot, and right now the parser can't handle trailing
commas, so I should think about that too, which if done right might
actually be the kind of change that makes the code even simpler than
when half as many features were in.