add array store to the interpreter
This is exciting! I can now initialise arrays, which is very satisfying.
From here I can build all the little features that make arrays really
work.

In order to get it together I also had to fix one small thing, which was
correcting the operation flags when pattern matching an array into
a new global/local. I also had to extract the memory modes of arg1 and
arg2, to work around the lack of type information in the runtime. I am
not sure if that is a good idea or not! Now that I have the modes, I
started checking that OP_MOV copies of arrays actually are arrays, since
I can. All other operations go unchecked, but could be added at some
point.