pull out some functions from expression parsing
I think these functions are still 'constraint bearing', but this is very
important structure to make explicit. As noted in the TODO comment for
this change, since these operations are transitions in a state machine,
the in-line structure actually doesn't communicate anything anyway, so
pulling some transitions out into functions and not others, actually
communicates a *lot*. I am not sure I have fully worked out the
theoretical conclusions here, but this change on its own is very nice,
so that is good.

Also fixed some warnings that were showing up on gcc, since I did this
on a Linux machine.