execute array index and concatenation
This is getting pretty nice, there is a reflection-based copy function
that all the array operations use now, that could have many more uses
going forwards.

A lot of little type propogation bugs showed up as soon as I tried to
print the results, and the result of concatenation wasn't an array! How
weird! Turns out I was basically ignoring type inference, and so
everything was defaulting back to the arithmetic default, int64.

Also now that I infer types after array construction, indexing, and
concatenation, there is just memory sharing going on all over the place.
I think that I want these types to be garbage collected, possibly as
shared_buff_headers??? Long term that might be the only way, if I want
to expose types to the user, or worse, allow the user to run arbitrary
code to generate types.