read from file iteratively
This means there is now a lot more data I am leaking, from all these
tokens. I will deal with it one day.

You can now write code from stdin as well, that is nice. I want to make
it so that if a full expression is encountered at a line ending, then it
just prints that expression.

It doesn't print final results on Windows due to some race condition
between stdin being closed and the program being killed, but it seems to
be working roughly the same still.