extend RPN into "pattern commands"
It was getting kind of silly working out how to encode the AST as tokens
plus "is_postfix" booleans, it's time to just make an AST, but keep it
serialized/linear the way the RPN representation was! Now we have
something fairly nice to work with, we'll see if adding struct literals
is also nice now... Once I work out how structs are supposed to work. :)