increment arrays in structs
The first part of this, writing arrays to structs in the first place, I
had working yesterday, but forgot to commit. Today I added the next
piece, incrementing arrays inside structs when the structs are copied.
Things are starting to move to a static destruction model, rather than
the current runtime-variable-unbinding model, but that would be a big
change.