remove memory modes
Instead instructions that manipulate arrays detect whether they are
acting on temporaries or not, similar to how struct literals don't
generate increment instructions if they are just being written
somewhere.

For now we just leak temporaries/locals at the end of statements and on
function return, since this change made a mess of the diff enough
already.