call/return using offset rather than count
Now that temporaries are manually destroyed just like locals, they don't
get popped from the variable stack, which means the top of the variable
stack is not necessarily the thing we want to use to call/return, so
functions called on complex expressions were going haywire. Now they
still do, but in a slightly different way, because we still have no way
of calculating locals_count correctly. If a function allocates enough
locals then it can avoid the problem now, by making so many locals that
the overestimation of the runtime becomes accurate after all... Anyway,
it seems like it is time to get rid of the local vs temporary
distinction altogether, but I will do that in another commit, since it
is a rather sweeping change, and since I also have a thing for
tuple/record types that I still need to commit as well.