parse tuple/record args
It kind of works as is, although passing temporaries into functions or
returning locals/temporaries out of functions will lead to weird
problems.

It also had an issue where the data stack was being freed by the callee,
even when the caller owned the struct, but I fixed that. It still
decrements any arrays inside of the structs that get passed in, which is
probably incorrect behaviour? I don't know.